feat: add systray integration for PR display and management
